Add FishingTime Config-Issues
Issue with file not being read. Value falls back to constructor value of 3.0, FishingTime.cfg is created correctly, and 1 value is wrote to it.

I think this is because it isn't made a ref. It's not a pointer so it gets garbage collected.

I plan on looking more into this. I have ran out of time this morning.