Prepare Action Rework
changed the min and max from being a flot to an int so it will more reliably use all numbers in the config. Also fixed the recipe animation that bohemia broke for at least our fillet knives. Animation still broken for base game knives. Also updated the config to make sure everything is an integer instead of a float.